refactor(gateway): extract response-normalization helpers from run.py (slice 16 of #54962)
